Fixed demo UI using new FDC3 UI Constants
