+++ toolbox/fdc3-for-web/demo/src/client/da/util.ts
@@ -2,15 +2,28 @@ import { InstanceID } from "@kite9/fdc3-web-impl"
 import { Socket } from "socket.io-client"
 import { FDC3_APP_EVENT, FDC3_DA_EVENT } from "../../message-types"
 
+export enum UI { DEFAULT, DEMO }
+
+export const UI_URLS = {
+    [UI.DEMO]: {
+        intentResolverUrl: window.location.origin + "/static/da/intent-resolver.html",
+        channelSelectorUrl: window.location.origin + "/static/da/channel-selector.html",
+    },
+    [UI.DEFAULT]: {
+        // TODO: REPLACE WITH FDC3.FINOS.ORG URLS AFTER GO-LIVE
+        intentResolverUrl: "http://localhost:4002/intent-resolver.html",
+        channelSelectorUrl: "http://localhost:4002/channel-selector.html",
+    }
+}
 
 export function link(socket: Socket, channel: MessageChannel, source: InstanceID) {
     socket.on(FDC3_DA_EVENT, (data: any, to: InstanceID) => {
-        console.log(`DA Sent ${JSON.stringify(data)} from socket`)
+        //console.log(`DA Sent ${JSON.stringify(data)} from socket`)
         channel.port2.postMessage(data)
     })
 
     channel.port2.onmessage = function (event) {
-        console.log(`App Sent ${JSON.stringify(event.data)} from message port`)
+        //console.log(`App Sent ${JSON.stringify(event.data)} from message port`)
         socket.emit(FDC3_APP_EVENT, event.data, source)
     }
 }
